+++ src/payloads/set_game_score_inline.rs
@@ -13,7 +13,7 @@ impl_payload! {
     pub SetGameScoreInline (SetGameScoreInlineSetters) => Message {
         required {
             /// User identifier
-            pub user_id: u32,
+            pub user_id: i64,
             /// New score
             pub score: u64,
             /// Identifier of the inline message
